refactor(engine): disconnecting framework from faux-shadow
