+++ packages/lwc-engine/src/faux-shadow/shadow-root.ts
@@ -1,13 +1,13 @@
 import assert from "../shared/assert";
 import { isNull, create, assign, isUndefined, toString, getOwnPropertyDescriptor, ArrayReduce, } from "../shared/language";
-import { getNodeKey } from "../framework/vm";
 import { addShadowRootEventListener, removeShadowRootEventListener } from "./events";
 import { shadowRootQuerySelector, shadowRootQuerySelectorAll, shadowRootChildNodes, getPatchedCustomElement, isNodeOwnedBy } from "./traverse";
 import { getInternalField, setInternalField, createFieldName } from "../shared/fields";
 import { getInnerHTML } from "../3rdparty/polymer/inner-html";
 import { getTextContent } from "../3rdparty/polymer/text-content";
-import { compareDocumentPosition, DOCUMENT_POSITION_CONTAINED_BY } from "./node";
-import { ElementAOMPropertyNames } from "../framework/attributes";
+import { compareDocumentPosition, DOCUMENT_POSITION_CONTAINED_BY, getNodeKey } from "./node";
+// it is ok to import from the polyfill since they always go hand-to-hand anyways.
+import { ElementPrototypeAriaPropertyNames } from "../polyfills/aria-properties/polyfill";
 import { unwrap } from "./traverse-membrane";
 
 let ArtificialShadowRootPrototype;
@@ -32,7 +32,7 @@ export function getShadowRoot(elm: HTMLElement): ShadowRoot {
 
 // Synthetic creation of all AOM property descriptors for Shadow Roots
 function createShadowRootAOMDescriptorMap(): PropertyDescriptorMap {
-    return ArrayReduce.call(ElementAOMPropertyNames, (seed: PropertyDescriptorMap, propName: string) => {
+    return ArrayReduce.call(ElementPrototypeAriaPropertyNames, (seed: PropertyDescriptorMap, propName: string) => {
         let descriptor: PropertyDescriptor | undefined;
         if (isNativeShadowRootAvailable) {
             descriptor = getOwnPropertyDescriptor((window as any).ShadowRoot.prototype, propName);
