+++ packages/lwc-engine/src/shared/fields.ts
@@ -1,3 +1,5 @@
+import { defineProperty } from "./language";
+
 /**
  * In IE11, symbols that we plan to apply everywhere are expensive
  * due to the nature of the symbol polyfill. This method abstract the
@@ -10,8 +12,19 @@ export function createFieldName(key: string): symbol {
 }
 
 export function setInternalField(o: object, fieldName: symbol, value: any) {
-    // TODO: improve this to use defineProperty(o, fieldName, { value }) or a WeakMap
-    o[fieldName] = value;
+    // TODO: improve this to use  or a WeakMap
+    if (process.env.NODE_ENV !== 'production') {
+        // in dev-mode, we are more restrictive about what you can do with the internal fields
+        defineProperty(o, fieldName, {
+            value,
+            enumerable: true,
+            configurable: false,
+            writable: false,
+        });
+    } else {
+        // in prod, for better perf, we just let it roll
+        o[fieldName] = value;
+    }
 }
 
 export function getInternalField(o: object, fieldName: symbol): any {
